refactor(emotion-carousel-choice-navigation)
