+++ app/screens/Reusable/Acknowledgement/AcknowledgementSuccessivelyDoubleButton/AcknowledgementSuccessivelyDoubleButton.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { ImageBackground, SafeAreaView, TouchableOpacity } from 'react-native';
 
@@ -8,17 +8,11 @@ import {
   usePositiveNavigateTo,
   useParsedJSXTextNickname,
   useRenderQuestHeader,
-  useAppSelector,
-  useAppDispatch,
   useNegativeNavigateTo,
-  useNavigateNextQuest,
 } from '../../../../hooks';
 import { generalStyles } from '../../../../utils/styles';
 import { IAcknowledgementSuccessivelyDoubleButtonScreenProps } from './AcknowledgementSuccessivelyDoubleButton.types';
 import { styles } from './AcknowledgementSuccessivelyDoubleButton.styles';
-import { cacheSlice } from '../../../../redux/slices';
-import { THE_CHARM_OF_THE_MIRROR_ID } from '../../../../constants/quest';
-import { EMOTION_BUTTON_LIST } from '../../EmotionSelection';
 import { CHARMS_SVG } from '../../../../assets/svg';
 
 export const AcknowledgementSuccessivelyDoubleButtonScreen: React.FC<IAcknowledgementSuccessivelyDoubleButtonScreenProps> =
@@ -38,21 +32,8 @@ export const AcknowledgementSuccessivelyDoubleButtonScreen: React.FC<IAcknowledg
     } = route.params.data;
 
     const { t } = useTranslation();
-    const dispatch = useAppDispatch();
-    const navigateToNextQuest = useNavigateNextQuest();
     const negativeNavigate = useNegativeNavigateTo(negativeNavigatesTo, true);
-    const positiveNavigate = usePositiveNavigateTo(positiveNavigatesTo);
-
-    const currentQuestLineId = useAppSelector(
-      state => state.quest.currentQuestLine?.id,
-    );
-    const completedEmotionsCount = useAppSelector(
-      state => state.cache.emotions.completed.length,
-    );
-
-    const isAllEmotionsCompleted =
-      completedEmotionsCount >= EMOTION_BUTTON_LIST.length - 1;
-    const isMirrorCharm = currentQuestLineId === THE_CHARM_OF_THE_MIRROR_ID;
+    const onSubmit = usePositiveNavigateTo(positiveNavigatesTo);
 
     const Title = useParsedJSXTextNickname({
       text: title,
@@ -65,32 +46,6 @@ export const AcknowledgementSuccessivelyDoubleButtonScreen: React.FC<IAcknowledg
       crossHeader: crossHeader ?? false,
       escapeMenuAlternativeNavigateTo,
     });
-
-    const onSubmit = useCallback(() => {
-      if (isMirrorCharm) {
-        dispatch(cacheSlice.actions.completeSelectedEmotion());
-
-        if (isAllEmotionsCompleted) {
-          navigateToNextQuest();
-          dispatch(cacheSlice.actions.clearEmotions());
-          return;
-        }
-      }
-      positiveNavigate();
-    }, [
-      dispatch,
-      isAllEmotionsCompleted,
-      isMirrorCharm,
-      navigateToNextQuest,
-      positiveNavigate,
-    ]);
-
-    const onBottomButtonPress = useCallback(() => {
-      if (isMirrorCharm) {
-        dispatch(cacheSlice.actions.clearEmotions());
-      }
-      negativeNavigate();
-    }, [dispatch, isMirrorCharm, negativeNavigate]);
 
     const Icon = image && CHARMS_SVG[image];
 
@@ -114,7 +69,7 @@ export const AcknowledgementSuccessivelyDoubleButtonScreen: React.FC<IAcknowledg
             </ExtendedText>
           </BottomButtonView>
           <TouchableOpacity
-            onPress={onBottomButtonPress}
+            onPress={negativeNavigate}
             style={styles.bottomButton}>
             <ExtendedText preset="secondary-text" style={styles.bottomButton}>
               {tellMoreTitle}
