+++ app/screens/Reusable/EmotionCarouselChoice/EmotionCarouselChoice.tsx
@@ -1,19 +1,23 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import { ImageBackground, SafeAreaView, View } from 'react-native';
 import { useTranslation } from 'react-i18next';
 
 import { IEmotionCarouselChoiceScreenProps } from './EmotionCarouselChoice.types';
 import { styles } from './EmotionCarouselChoice.styles';
 import {
+  useAppDispatch,
   useAppSelector,
   useNavigateNextQuest,
   useParsedJSXTextNickname,
+  usePositiveNavigateTo,
   useRenderQuestHeader,
 } from '../../../hooks';
 import { CHARMS_BACKGROUNDS } from '../../../assets';
 import { generalStyles } from '../../../utils/styles';
 import { BottomButtonView, ExtendedText } from '../../../components';
 import { EMOTIONS_CAROUSEL_SVG } from '../../../assets/svg';
+import { EMOTION_BUTTON_LIST } from '../EmotionSelection';
+import { cacheSlice } from '../../../redux/slices';
 
 export const EmotionCarouselChoiceScreen: React.FC<IEmotionCarouselChoiceScreenProps> =
   ({ route }) => {
@@ -24,13 +28,33 @@ export const EmotionCarouselChoiceScreen: React.FC<IEmotionCarouselChoiceScreenP
       backgroundImage,
       crossHeader,
       titleHasNickname,
+      positiveNavigatesTo,
       escapeMenuAlternativeNavigateTo,
     } = route.params.data;
 
     const { t } = useTranslation();
-    const { emotionItem } = useAppSelector(state => state.cache);
+    const dispatch = useAppDispatch();
+    const navigateToNextQuest = useNavigateNextQuest();
+    const positiveNavigate = usePositiveNavigateTo(positiveNavigatesTo);
+    const { emotionItem, emotions } = useAppSelector(state => state.cache);
+    const completedEmotionsCount = emotions.completed.length;
+    const isAllEmotionsCompleted =
+      completedEmotionsCount >= EMOTION_BUTTON_LIST.length - 1;
 
-    const onSubmit = useNavigateNextQuest();
+    const onSubmit = useCallback(() => {
+      dispatch(cacheSlice.actions.completeSelectedEmotion());
+      if (isAllEmotionsCompleted) {
+        positiveNavigate();
+        dispatch(cacheSlice.actions.clearEmotions());
+        return;
+      }
+      navigateToNextQuest();
+    }, [
+      dispatch,
+      isAllEmotionsCompleted,
+      navigateToNextQuest,
+      positiveNavigate,
+    ]);
 
     const Title = useParsedJSXTextNickname({
       text: title,
